"Superstore" Retail Store Analysis and Analytics
In this project, I performed exploratory data analyses of a retail company named "Superstore", using python programming. I also used machine learning tools to cluster its customers and categorize their data into high-paying clients and likely irregular or one-time purchasers. This data is crucial in helping firms make market-targeting decisions and measure their outcomes. It also helps in customer satisfaction and retention as it is a way of ensuring that customers are satisfied with the product/value that they spend on.