Telco Customer Churn Analysis
A Telecommunication company in ABC has recently experienced a decrease in customer retention that has signed up for its services and wants to find out ways to improve their customer retention.

As the data analyst in the company, I have been tasked to work on this project. I have identified the all relevant customer data I needed to generate insights. The task is to answer these questions:
•	What are the key drivers of churn and what influences them?
•	What are the appropriate churn initiatives that should be launched to address them?